Removed duplicate types & moved to separate file
