+++ app/api/purchase/list/route.ts
@@ -4,22 +4,7 @@ import { authOptions } from "../../auth/[...nextauth]/route";
 import { collection, getDocs } from "firebase/firestore";
 import { db } from "@/lib/firebaseConfig";
 import PurchaseEntry from "@/models/purchase";
-
-interface Items {
-  itemName: string;
-  price: string;
-  quantity: string;
-}
-
-export interface Purchase {
-  id: string;
-  date: string;
-  totalAmount: number;
-  userId: string;
-  vendorId: string;
-  vendorName: string;
-  items: Items[];
-}
+import { Purchase } from "@/types/types";
 
 export async function POST(req: NextRequest) {
   const session = await getServerSession(authOptions);
